requirements.txt
streamlit
plotly
kaleido
fpdf
python-dotenv
streamlit
matplotlib
numpy
fpdf
python-dotenv
